Cleanup macro PCP checks
