+++ tests/neg/splice-in-top-level-splice-1.scala
@@ -1,7 +1,7 @@
 import scala.quoted._
 
 object Foo {
-  inline def foo(): Int = ${bar($x)} // error
+  inline def foo(): Int = ${bar(${x})} // error
   def x: Expr[Int] = '{1}
   def bar(i: Int): Expr[Int] = i.toExpr
 }
